+++ examples/resources/fmc_intrusion_rule_group/resource.tf
@@ -1,0 +1,4 @@
+resource "fmc_intrusion_rule_group" "example" {
+  name        = "my_intrusion_rule_group"
+  description = "My IPS Rule Group"
+}